# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Toya Amerta API",
        "description": "REST API untuk aplikasi mobile (Flutter) Sistem Informasi PDAM Desa Toya Amerta. Semua response menggunakan envelope { success, message, data, meta }.",
        "contact": {
            "name": "Toya Amerta Dev Team",
            "email": "pt.andisuartika@gmail.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/petugas/tagihan": {
            "get": {
                "tags": [
                    "Petugas - Pembayaran"
                ],
                "summary": "Daftar tagihan belum bayar",
                "operationId": "19060f7e48a3e2c874375e09fe03d9f3",
                "parameters": [
                    {
                        "name": "zone_id",
                        "in": "query",
                        "description": "Filter zona",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "year",
                        "in": "query",
                        "description": "Filter tahun periode",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 2026
                        }
                    },
                    {
                        "name": "month",
                        "in": "query",
                        "description": "Filter bulan periode",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 6
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "OK"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 55
                                                    },
                                                    "customer_id": {
                                                        "type": "integer",
                                                        "example": 5
                                                    },
                                                    "customer_name": {
                                                        "type": "string",
                                                        "example": "Wayan Karya"
                                                    },
                                                    "customer_number": {
                                                        "type": "string",
                                                        "example": "PLG-0005"
                                                    },
                                                    "zone": {
                                                        "type": "string",
                                                        "example": "Zona A"
                                                    },
                                                    "period_year": {
                                                        "type": "integer",
                                                        "example": 2026
                                                    },
                                                    "period_month": {
                                                        "type": "integer",
                                                        "example": 6
                                                    },
                                                    "period_label": {
                                                        "type": "string",
                                                        "example": "Juni 2026"
                                                    },
                                                    "total_amount": {
                                                        "type": "number",
                                                        "format": "float",
                                                        "example": 31000
                                                    },
                                                    "remaining_amount": {
                                                        "type": "number",
                                                        "format": "float",
                                                        "example": 31000
                                                    },
                                                    "payment_status": {
                                                        "type": "string",
                                                        "example": "belum_bayar",
                                                        "enum": [
                                                            "belum_bayar",
                                                            "sebagian"
                                                        ]
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "meta": {
                                            "properties": {
                                                "total": {
                                                    "type": "integer",
                                                    "example": 1
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UnauthorizedResponse"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Role tidak diizinkan",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ForbiddenResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/petugas/payments": {
            "post": {
                "tags": [
                    "Petugas - Pembayaran"
                ],
                "summary": "Konfirmasi pembayaran",
                "description": "Mendukung pembayaran sebagian (cicilan). amount_paid tidak boleh melebihi sisa tagihan. Setelah konfirmasi, transaksi kas masuk otomatis tercatat.",
                "operationId": "d437ea0fea191186ade2675b206a5ebe",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "water_reading_id",
                                    "amount_paid",
                                    "payment_date",
                                    "payment_method"
                                ],
                                "properties": {
                                    "water_reading_id": {
                                        "type": "integer",
                                        "example": 55
                                    },
                                    "amount_paid": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 31000
                                    },
                                    "payment_date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2026-06-09"
                                    },
                                    "payment_method": {
                                        "type": "string",
                                        "example": "tunai",
                                        "enum": [
                                            "tunai",
                                            "transfer",
                                            "qris"
                                        ]
                                    },
                                    "notes": {
                                        "type": "string",
                                        "example": "Bayar lunas",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Pembayaran berhasil dikonfirmasi",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Pembayaran berhasil dikonfirmasi."
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/PaymentRecord"
                                        },
                                        "meta": {
                                            "type": "object",
                                            "example": null,
                                            "nullable": true
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UnauthorizedResponse"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Role tidak diizinkan",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ForbiddenResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validasi gagal / sudah lunas / melebihi sisa tagihan",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/petugas/payments/{id}": {
            "get": {
                "tags": [
                    "Petugas - Pembayaran"
                ],
                "summary": "Detail pembayaran",
                "operationId": "aef57d6b8c4ff91f416ac9d71acf3fd3",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID payment record",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "OK"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/PaymentRecord"
                                        },
                                        "meta": {
                                            "type": "object",
                                            "example": null,
                                            "nullable": true
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UnauthorizedResponse"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Role tidak diizinkan",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ForbiddenResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Tidak ditemukan",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/NotFoundResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "ValidationErrorResponse": {
                "title": "Validation Error Response",
                "properties": {
                    "success": {
                        "type": "boolean",
                        "example": false
                    },
                    "message": {
                        "type": "string",
                        "example": "Data tidak valid."
                    },
                    "data": {
                        "type": "object",
                        "example": {
                            "field": [
                                "Pesan error validasi."
                            ]
                        }
                    },
                    "meta": {
                        "type": "object",
                        "example": null,
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "UnauthorizedResponse": {
                "title": "Unauthenticated Response",
                "properties": {
                    "success": {
                        "type": "boolean",
                        "example": false
                    },
                    "message": {
                        "type": "string",
                        "example": "Unauthenticated. Token tidak valid atau sudah kedaluwarsa."
                    },
                    "data": {
                        "type": "object",
                        "example": null,
                        "nullable": true
                    },
                    "meta": {
                        "type": "object",
                        "example": null,
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "ForbiddenResponse": {
                "title": "Forbidden Response",
                "properties": {
                    "success": {
                        "type": "boolean",
                        "example": false
                    },
                    "message": {
                        "type": "string",
                        "example": "Akses ditolak. Role Anda tidak memiliki izin."
                    },
                    "data": {
                        "type": "object",
                        "example": null,
                        "nullable": true
                    },
                    "meta": {
                        "type": "object",
                        "example": null,
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
            "PaymentRecord": {
                "title": "Payment Record",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 30
                    },
                    "receipt_number": {
                        "type": "string",
                        "example": "KWT-20260609-0001"
                    },
                    "water_reading_id": {
                        "type": "integer",
                        "example": 55
                    },
                    "customer_id": {
                        "type": "integer",
                        "example": 5
                    },
                    "period_label": {
                        "type": "string",
                        "example": "Juni 2026"
                    },
                    "amount_paid": {
                        "type": "number",
                        "format": "float",
                        "example": 31000
                    },
                    "payment_date": {
                        "type": "string",
                        "format": "date",
                        "example": "2026-06-09"
                    },
                    "payment_method": {
                        "type": "string",
                        "example": "tunai",
                        "enum": [
                            "tunai",
                            "transfer",
                            "qris"
                        ]
                    },
                    "status": {
                        "type": "string",
                        "example": "lunas",
                        "enum": [
                            "sebagian",
                            "lunas"
                        ]
                    },
                    "recorded_by": {
                        "type": "string",
                        "example": "Budi Santoso"
                    },
                    "notes": {
                        "type": "string",
                        "example": "Bayar lunas",
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
            "NotFoundResponse": {
                "title": "Not Found Response",
                "properties": {
                    "success": {
                        "type": "boolean",
                        "example": false
                    },
                    "message": {
                        "type": "string",
                        "example": "Data tidak ditemukan."
                    },
                    "data": {
                        "type": "object",
                        "example": null,
                        "nullable": true
                    },
                    "meta": {
                        "type": "object",
                        "example": null,
                        "nullable": true
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Gunakan token yang didapat dari endpoint /auth/login. Format header: Authorization: Bearer {token}",
                "bearerFormat": "token",
                "scheme": "bearer"
            }
        }
    },
...
}